backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pycryptodome
python-dotenv
cryptography
hvac